refactor(a2a): merge probeClient interface into interop_probe_clients_test.go
interop_client_test.go only held the probeClient interface and newClientFn
type — both belong alongside their implementations. Move them into
interop_probe_clients_test.go and delete the file.
